refactor: extract floating text service
