+++ internal/src/core/ipc/IpcBridge.cpp
@@ -6,6 +6,7 @@
 #include "IpcMessages.h"
 #include "IpcSession.h"
 #include "FeatureState.h"
+#include "FloatingTextService.h"
 
 // Debug logging
 
@@ -22,19 +23,6 @@
 
 static const char* PipeName() { return BUILD_PIPE_NAME; }
 static const DWORD PIPE_BUFFER_SIZE = 65536;
-
-static std::mutex s_pluginFloatingTextMutex;
-static char s_pluginFloatingText[128] = {};
-static bool s_pluginFloatingTextPending = false;
-
-struct UnityNullableColor32Abi {
-    bool hasValue;
-    uint8_t padding[3];
-    uint32_t rgba;
-};
-
-static_assert(sizeof(UnityNullableColor32Abi) == 8, "Nullable<Color32> ABI must be 8 bytes");
-constexpr uint32_t PackColor32(uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255) { return uint32_t(r) | (uint32_t(g) << 8) | (uint32_t(b) << 16) | (uint32_t(a) << 24); }
 
 // Tile map API
 
@@ -263,61 +251,13 @@ void    IpcBridge_SetSkinOverride(bool enabled, int skinId)   { FeatureState::Se
 int32_t IpcBridge_GetClientDefense()                          { return FeatureState::GetClientDefense(); }
 int32_t IpcBridge_GetClientClassType()                        { return FeatureState::GetClientClassType(); }
 
-// Floating text and render-thread apply entry
-
-static void ApplyPluginFloatingTextFeatureState()
-{
-    char text[128] = {};
-    {
-        std::lock_guard<std::mutex> lk(s_pluginFloatingTextMutex);
-        if (!s_pluginFloatingTextPending) return;
-        strncpy_s(text, sizeof(text), s_pluginFloatingText, _TRUNCATE);
-    }
-
-    Il2CppClass* klass = Resolver::FindClassLoose("MapObjectUIManager");
-    const MethodInfo* mi = klass ? il2cpp_class_get_method_from_name(klass, "ShowFloatingText", 6) : nullptr;
-    app::MapObjectUIManager* localMgr = nullptr;
-    void* local = GameState::GetLocalPtr();
-
-    Resolver::Protection::safe_call([&]() {
-        auto* localView = *reinterpret_cast<app::ViewHandler**>(reinterpret_cast<uintptr_t>(local) + RuntimeOffsets::KJ_ViewHandler);
-        if (localView) localMgr = localView->fields.GUIManager;
-    });
-
-    void* receiver = localMgr;
-    if (!receiver && klass) { auto objs = Resolver::FindObjectsByType(klass); if (!objs.empty()) receiver = objs[0]; }
-    if (!receiver || !mi || !mi->methodPointer) return;
-
-    using Fn = void(*)(void*, app::DGKAANOAENH__Enum, app::String*, UnityNullableColor32Abi, float, float, float, const MethodInfo*);
-    auto showFloatingText = reinterpret_cast<Fn>(mi->methodPointer);
-    const bool off = strstr(text, "Disabled") != nullptr;
-
-    UnityNullableColor32Abi col{};
-    col.hasValue = true; col.rgba = off ? PackColor32(255, 0, 25) : PackColor32(32, 220, 0);
-
-    static void* s_primedReceiver = nullptr;
-    if (s_primedReceiver != receiver) {
-        app::String* emptyText = reinterpret_cast<app::String*>(il2cpp_string_new(""));
-        const bool primeOk = Resolver::Protection::safe_call([&]() { for (int i = 0; i < 12; ++i) showFloatingText(receiver, app::DGKAANOAENH__Enum::Xp, emptyText, col, 0.f, 0.f, 0.f, mi); });
-        if (primeOk) s_primedReceiver = receiver;
-        return;
-    }
-
-    app::String* ilText = reinterpret_cast<app::String*>(il2cpp_string_new(text));
-    const bool ok = Resolver::Protection::safe_call([&]() { showFloatingText(receiver, app::DGKAANOAENH__Enum::Xp, ilText, col, 0.f, 0.f, 0.f, mi); });
-    if (ok) {
-        std::lock_guard<std::mutex> lk(s_pluginFloatingTextMutex);
-        if (strcmp(s_pluginFloatingText, text) == 0) s_pluginFloatingTextPending = false;
-    }
-}
-
 void IpcBridge_ApplyFeatureOverrides()
 {
     ApplyPlayerNoclipFeatureState();
     if (GameState::GetLocalPtr() == nullptr) return;
     if (GameState::GetWorldMgr() == nullptr) return;
     ApplyAutoAimFeatureState(); ApplyProjectileNoclipFeatureState(); ApplyAutoDodgeFeatureState(); ApplyAutoAbilityFeatureState();
-    ApplyWalkTargetFeatureState(); ApplyCameraFeatureState(); ApplyPluginFloatingTextFeatureState();
+    ApplyWalkTargetFeatureState(); ApplyCameraFeatureState(); FloatingTextService::ApplyPendingPluginText();
 }
 
 static std::atomic<bool> s_shutdown{false};
@@ -474,13 +414,6 @@ static bool ParseSetFeatureCommand(char* json, const char* seqStr, const char* m
     return true;
 }
 
-static void QueuePluginFloatingText(const char* text)
-{
-    std::lock_guard<std::mutex> lk(s_pluginFloatingTextMutex);
-    strncpy_s(s_pluginFloatingText, sizeof(s_pluginFloatingText), text ? text : "", _TRUNCATE);
-    s_pluginFloatingTextPending = true;
-}
-
 static bool DispatchTileCommand(const char* type, char* json, const char* seqStr, const char* macHex)
 {
     if (strcmp(type, "clearTiles") == 0) {
@@ -542,7 +475,7 @@ static bool ApplyCoreFeature(const FeatureCommand& f)
         FH_FLOAT("autoAbilityMpPct", IpcBridge_SetAutoAbilityMpPct),
         FH_INT("autoAbilityWizardMode", IpcBridge_SetAutoAbilityWizardMode),
         FH_INT("targetFrameRate", FpsSetter::SetTargetFps),
-        FH_TEXT("showPluginFloatingText", QueuePluginFloatingText)
+        FH_TEXT("showPluginFloatingText", FloatingTextService::QueuePluginText)
     };
     return ApplyFeatureTable(f, h, sizeof(h) / sizeof(h[0]));
 }
