+++ internal/src/features/visuals/FloatingTextService.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace FloatingTextService {
+
+void QueuePluginText(const char* text);
+void ApplyPendingPluginText();
+
+} // namespace FloatingTextService
